+++ src/pages/Homework/CreateHomework.tsx
@@ -2,48 +2,26 @@ import { ChangeEvent, useState } from 'react';
 import 'react-datepicker/dist/react-datepicker.css';
 import { useParams } from 'react-router-dom';
 import { uploadHomework } from '../../api/homework.api';
-import AddIcon from '../../assets/images/File_dock_add.svg?react';
 import LongButton from '../../components/LongButton';
 import Calendar from '../../assets/images/calendarGray.svg?react';
 import { MdKeyboardArrowRight } from 'react-icons/md';
 import DatePicker from '../../components/DatePicker';
 import Modal from '../../components/Modal/Modal';
+import AddFile from '../../components/AddFile';
 
 const CreateHomework = () => {
   const { roomId } = useParams<{ roomId: string }>();
   const [fileList, setFileList] = useState<File[]>([]); // 파일 이름 목록
-  const [isActive, setIsActive] = useState<boolean>(false); // 드래그 활성화 상태
   const [deadline, setDeadline] = useState<string>(''); // 선택된 숙제 마감 날짜
   const [desc, setDesc] = useState<string>(''); // 숙제 설명
   const [showDatePicker, setShowDatePicker] = useState(false);
 
-  const handleDragStart = () => setIsActive(true);
-  const handleDragEnd = () => setIsActive(false);
-
-  const handleFileChange = (event: ChangeEvent<HTMLInputElement>) => {
-    const files = event.target.files;
-    if (files) {
-      const fileArray = Array.from(files);
-      setFileList([...fileList, ...fileArray]);
-    }
-  };
-
-  const handleDrop = (event: React.DragEvent<HTMLLabelElement>) => {
-    event.preventDefault();
-    setIsActive(false);
-    const fileList = event.dataTransfer.files;
-    const fileArray = Array.from(fileList);
-    setFileList([...fileList, ...fileArray]);
-  };
-
+  // 설명 추가
   const handleDesc = (event: ChangeEvent<HTMLInputElement>) => {
     setDesc(event.target.value);
   };
 
-  const handleDeleteFile = (index: number) => {
-    setFileList((prev) => prev.filter((_, i) => i !== index));
-  };
-
+  // 마감날짜와 설명 있는지 확인
   const validateForm = (): boolean => {
     if (!deadline) {
       alert('마감 날짜를 선택해주세요');
@@ -109,41 +87,8 @@ const CreateHomework = () => {
             onChange={handleDesc}
           ></input>
         </div>
-        <div
-          className={`border-[1px] h-[200px] border-dashed rounded-lg py-10 flex flex-col items-center justify-center cursor-pointer ${
-            isActive ? 'border-primary_500 bg-primary_50' : 'border-[#D4D4D8] bg-[#FAFAFA]'
-          }`}
-        >
-          <label
-            onDragEnter={handleDragStart}
-            onDragLeave={handleDragEnd}
-            onDrop={handleDrop}
-            onDragOver={(e) => e.preventDefault()}
-            className="w-full flex items-center justify-center"
-          >
-            <input type="file" className="hidden" multiple onChange={handleFileChange} />
-            <div className="justify-center items-center flex flex-col gap-1">
-              <AddIcon className="p-2 w-12 h-12 bg-gray-100 rounded-full" />
-              <p className="text-gray-500 text-[12px] leading-[22px]">파일 첨부하기</p>
-            </div>
-          </label>
-        </div>
+        <AddFile setFileList={setFileList} fileList={fileList} />
       </div>
-      {fileList.length > 0 && (
-        <div className="mt-4 text-center text-gray-700">
-          <p>업로드된 파일 목록:</p>
-          <ul>
-            {fileList.map((file, index) => (
-              <li key={index} className="font-bold">
-                {file.name}
-                <button onClick={() => handleDeleteFile(index)} className="text-red-500 hover:underline ml-4">
-                  삭제
-                </button>
-              </li>
-            ))}
-          </ul>
-        </div>
-      )}
       <div className="py-6 mt-auto">
         <LongButton
           enable={!!(fileList.length > 0 && desc.length > 0 && deadline.length > 0)}
